test: split agentic KPI tile IT into per-category classes
AgenticKpiTilesIT had grown to ~1600 lines covering every agentic
dashboard KPI tile, making it slow to navigate and hard to own. Split
it into five category-scoped integration tests — Execution, Token usage,
Reliability, Tool calls, and Duration — so each file maps to a single
report category and stays focused.

Rather than introduce a shared abstract superclass (which would force
inheritance just to share helpers), the common logic is extracted into
composition collaborators so each test depends only on what it uses:

- AgenticInstanceFixtures: static test-data builders. Placed in the
  io.camunda.optimize package so it can reach the protected static
  AbstractBrokerlessZeebeCCSMIT#completedInstance seam without subclassing.
- AgenticReportEvaluator: wraps ReportEvaluationService and exposes the
  evaluate* helpers the tests need; held as a field built in @BeforeEach.
- AgenticReportFilters: pure static builders for the additional report
  evaluation filters.
